refactor: create page component
