+++ src/components/Button/button.module.scss
@@ -1,11 +1,14 @@
 .button {
   border: none;
-  width: 100px;
+  width: fit-content;
   font-size: 20px;
+  height: 36px;
   border-radius: 10px;
   transition: 0.2s;
   color: #ffffff;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  box-shadow:
+    0 1px 3px rgba(0, 0, 0, 0.12),
+    0 1px 2px rgba(0, 0, 0, 0.24);
 
   &:hover {
     cursor: pointer;
@@ -25,6 +28,7 @@
 .error {
   background-color: #db4437;
   border: 1px solid #db4437;
+  align-self: end;
 
   &:hover {
     color: #db4437;
