+++ src/components/ErrorBoundary/error-boundary.module.scss
@@ -6,5 +6,7 @@
   padding: 20px;
   flex-direction: column;
   align-items: center;
-  box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+  box-shadow:
+    0 3px 6px rgba(0, 0, 0, 0.16),
+    0 3px 6px rgba(0, 0, 0, 0.23);
 }
