+++ src/components/Page/page.module.scss
@@ -1,0 +1,4 @@
+.page {
+  display: flex;
+  flex-direction: column;
+}
