+++ src/components/Result/result.module.scss
@@ -1,13 +1,15 @@
 .result {
   border-radius: 20px;
   padding: 20px;
-  box-shadow: 0 3px 6px rgba(0,0,0,0.16), 0 3px 6px rgba(0,0,0,0.23);
+  box-shadow:
+    0 3px 6px rgba(0, 0, 0, 0.16),
+    0 3px 6px rgba(0, 0, 0, 0.23);
   margin: 30px 0;
 }
 
 .table {
   border-collapse: collapse;
-  width: 100%
+  width: 100%;
 }
 
 .header-cell,
