+++ src/components/Search/search.module.scss
@@ -4,7 +4,9 @@
   gap: 20px;
   border-radius: 20px;
   padding: 20px;
-  box-shadow: 0 3px 6px rgba(0,0,0,0.16), 0 3px 6px rgba(0,0,0,0.23);
+  box-shadow:
+    0 3px 6px rgba(0, 0, 0, 0.16),
+    0 3px 6px rgba(0, 0, 0, 0.23);
 }
 
 .search-bar {
