+++ src/main.tsx
@@ -1,7 +1,6 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './App.tsx';
-import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary.tsx';
 
 const container = document.getElementById('root');
 if (!container) {
@@ -10,8 +9,6 @@ if (!container) {
 
 createRoot(container).render(
   <StrictMode>
-    <ErrorBoundary>
-      <App />
-    </ErrorBoundary>
+    <App />
   </StrictMode>
 );
